+++ app/routes/modelRouter.js
@@ -13,38 +13,45 @@ router.get('/upload-model', (req, res) => {
     });
 });
 
-router.post('/upload-model', upload_model.single('data-model'), (req, res) => {
+router.post('/upload-model', upload_model.single('data-model'), async (req, res) => {
     const extAccept = ['.pt', '.pth']
     const extname = path.extname(req.file.originalname).toLowerCase()
 
     if (extAccept.includes(extname)) {
-				fs.unlinkSync(path.join(__dirname, "../public/model/best.pt"));
-        fs.rename(
+        await fs.rename(
             path.join(__dirname, '../public/model/', req.file.originalname),
             path.join(__dirname, '../public/model/', 'best.pt'),
             (err) => {
                 if (err) throw err;
             }
         );
-        fs.unlinkSync(path.join(__dirname, "../../best.pt"));
+        if (await fs.existsSync(path.join(__dirname, "../../best.pt"))) {
+            await fs.unlinkSync(path.join(__dirname, "../../best.pt"));
+        }
+
         const path_src = path.join(__dirname, "../public/model/best.pt");
         const path_dst = path.join(__dirname, "../../best.pt");
-        fs.copyFile(path_src, path_dst, (err) => {
+        await fs.copyFile(path_src, path_dst, (err) => {
             if (err) {
                 console.log(err);
             }
             console.log("Copy file done.");
         });
-        const hasModel = true
+        const justAdd = true
         return res.render('upload-model', {
-            hasModel
+            justAdd
+        });
+    } else {
+        if (await fs.existsSync(path.join(__dirname, "../../best.pt"))) {
+            await fs.unlinkSync(path.join(__dirname, "../../best.pt"));
+        }
+        await fs.unlinkSync(path.join(__dirname, '../../', req.file.originalname));
+        const notSupport = true
+        return res.render('upload-model', {
+            notSupport
         });
     }
-    fs.unlinkSync(path.join(__dirname, '../../', req.file.originalname));
-    const notSupport = true
-    return res.render('upload-model', {
-        notSupport
-    });
+
 });
